feat: cc-0105-refactor the table header and pagination button logic
