--- src/constants/courtsTabelHeader.ts
@@ -1,1 +1,0 @@
-export const courtsTableHeader = ["Name", "Created At", "Updated At", "Detail", "Display"];
