+++ src/constants/tabelHeaders.ts
@@ -1,0 +1,17 @@
+export const COURTS_TABLE_HEADER = ["Name", "Created At", "Updated At", "Detail", "Display"];
+export const USERS_TABLE_HEADER = [
+  "Account ID",
+  "First Name",
+  "Last Name",
+  "Email",
+  "Account Type",
+];
+export const ADMIN_TABLE_HEADER = [
+  "Admin Name",
+  "Email",
+  "Created At",
+  "Updated At",
+  "Permission",
+  "Status",
+  "Operation",
+];
