--- src/constants/usersTableHeader.ts
@@ -1,7 +1,0 @@
-export const USERS_TABLE_HEADER = [
-  "Account ID",
-  "First Name",
-  "Last Name",
-  "Email",
-  "Account Type",
-];
